Move doRequest logic to separate file
